Loader system for loading the config file
